check if docker URI is valid
